Remove %current_scope variable.
This is likely the same as what is on irscope_removal but I noticed
TemporaryCurrentScopeVariable is not removed there yet.  This should be totally safe
for 9.2.10.0 since:
a) everything in interp and jit use current staticscope anyways
b) inliner already has special logic for target scopes to use an explicit Scope operand

As an extra commit I can document why I think this is safe and why the inliner will
not be broken (see a/b above). Plus I am jet-lagged and was having a hard time
concetrating on issue triage for 9.2.10.0...