Add a dummy JIT compiler to allow fully disabling
For enviroments like GraalVM native compile that can't support any
form of JVM bytecode JIT, we need to be able to completely disable
this code. For now the only options are the "tiered" JIT compiler
which can run both the full IR and JVM bytecode optimizations,
and this dummy JIT compiler that does nothing.