Begin to add formal "ephemeral instance" support
This feature has floated around for years as a way to pre-boot
background instances via tools like Nailgun or Drip. It will work
in a similar way with GraalVM native image, which can pre-boot and
freeze the heap for quick restarts. This commit begins to
formalize the process of constructing and reinitializing such an
ephemeral instance of Ruby.