Enlist expand_path logic in Dir.glob base handling
Dir.glob was treating any passed-in `base` path as a normal file
path, which cause it to treat URIs as a relative path. The logic
then proceeded to prepend the cwd to this URI, producing a
nonsense path.

This patch uses File.expand_path logic to acquire the base path,
which keeps it a URI and allows the rest of glob to handle it
as if the glob and base were expanded before calling Dir.glob.

The change means that the two following Dir.glob calls work the
same, fixing the issues discovered in #6060. It will very likely
also fix the issues reported in #6082 and #6083.

Fixes #6060
Fixes #6082
Fixes #6083