Revert "[refactor] minor - array-list size might be known"
This reverts commit ed72b695584909e3db384732c85643dcafbe337d.

This appears to be the source of a dramatic performance regression
in unbounded `caller` and `caller_locations` calls. Those paths
pass `1 << 24` as the maximum size for the resulting trace.
Previously, this maximum size was just used as an upper bound for
the stack gathering, but ed72b695 uses it as an initial size for
the ArrayList. This results in a 64KB array being allocated for
every such `caller` call, and a nearly 1000x slowdown as a result.

Note this was discovered via a backtrace from the neo4j
"active_graph" use of ActiveSupport instrumentation here:

https://github.com/neo4jrb/neo4j/blob/active_graph_rename/lib/active_graph/core/querable.rb#L59