Better test for Java 9+
Previous code looked for the Runtime.Version class added in Java 9
but failed to properly use a `$` for the inner class. This patch
changes to `java.lang.Module` which is a simple target and more
germane to the sorts of access checks we want to perform.

This fixes part of #6100