Tweaks to align java_import with other class-level methods
`java_import` has been defined on Object for many years, but after
all that time it seems it really should have been defined on
Module and toplevel only, similar to the visibility methods and
`define_method`. In addition, it uses unnecessary evals to query
and assign the constant, and if called against a non-Module it
always defines the resulting constant on Object. This leads to
peculiar behavior as reported in #6115.

This change does the following:

* Deprecate and warn for java_import called against arbitrary
  non-Module objects.
* Move java_import to Module and define a top-level version that
  invokes it against Object.
* Eliminate the eval and use const* methods to assign the value.

This should be a bit faster (no evals) and provide a path toward
removing Object#java_import.

All spec:ji passes without any warnings, so at least we are not
using Object#java_import there.