Ensure thread stays referenced until started
This is the simple fix for #6142. The primary flaw here was doing
the start call by traversing the weak reference in NativeThread,
which allowed a tiny window between the last access of the
`thread` variable and the starting of that thread, during which
GC could run and clear the weak reference. This commit also fixes
NativeThread.start to fail hard if this situation ever happens
again, so we'll at least get a clear error (though it should not
be possible for to happen anymore).