Remove these checks since they won't work with ID strings
These methods must receive "ID strings" (encoded as ISO-8859-1)
in order to properly populate the constant table, but the
`IdUtil.isConstant` logic does not work with such strings when
the decoded version has leading unicode characters. We remove the
checks here and need to audit the upstream callers to ensure they
are not passing value ID strings in.

See #6144 for ongoing work to audit constant table updates from
Java.