Use new symbol type classifier for constants.
This will be expanded and sped up a bit for ascii but
it uses MRI-friendly symbol type classification.  Now
when something wants to know if a symbol is a constant
it will just be a single enum equality check.

The downside of this is potentially eagerly calculating
this for all symbols made.  The vast majority of symbols
created will end up performing a check like this at
some point and in uncommon cases where that symbol is
used in the same way in a loop it will be an improvement.