Remove JVM shutdown hook for teardown
There's not a lot of context on the commit from 2013 that added
this, but I suspect it's not really correct behavior anymore. In
the process of reordering teardown (#6213) this logic began to
cause a spec hang for Kernel.exit! specs due to the teardown
deadlocking with the main thread:

* Main thread starts up and joins a thread that calls exit!. The
  java.lang.Thread.join call locks the thread object.
* Thread calls exit!, which calls System.exit (as opposed to the
  non-bang exit, which terminates the main thread and lets it tear
  things down safely).
* System exit calls this teardown hook, which attempts to shut
  terminate and join remaining threads.
* Because the teardown is not called from the main thread, the
  system deadlocks trying to shut down the child thread, which is
  locked in the main thread doing join.

The chain of deadlock is main => thread 1 join => system exit =>
thread 2 for teardown => try to kill and join thread 1. The
ultimate deadlock is due to a circular sequence of calls that
can't be interrupted (System.exit's wait on the teardown thread)
combined with the teardown happening from a non-main thread.

It's not clear why this started to hang after the reordering.
Moving the ThreadService teardown back to its former relative
position does not fix the hang, and logically I don't know why it
would... there should still be a deadlock if main is joining a
thread and the System exit hook tries to tear that thread down.