Output something to keep specs from timing out
As specs approach a ten minute run, we start to risk CI systems
terminating our builds for lack of output. This seems to be the
case recently with the :slow rubyspec suite, which has crept up
past 11 minutes total recently. With the previous change to
disable specdoc, it now fails to output anything for over ten
minutes and is terminated.

This commit changes the CI output from "summary" (which only
outputs at the end) to "file", which outputs a dot per spec file
executed. This will be a sufficient "heartbeat" for CI services
like Travis, but not so noisy like the full specdoc or a dot-
per-spec.

Note that the main :fast rubyspec suite is only just crossing
eight minutes, but without this change it could conceivably start
to fail in the future as well.

The MRI core suites are all running well over ten minutes, but the
tests in question are apparently chatty enough to keep the session
alive. They may become a problem later, but this is simple enough
to diagnose and remedy.