Allow public methods from non-public superclasses to bind
The existing logic will never bind any methods declared on a non-
public superclass, but this is incorrect behavior if the subclass
is public. A public subclass will expose public methods from any
superclass, regardless of the superclass's visibility, so we need
to also bind those methods.

There are four changes here:

* Allow methods to be gathered from superclasses if the superclass
  is public or the target class is public.
* Allow those methods to be bound even if there's no definition on
  the target class (no so-called "local method")
* Consider methods equivalent if parent method's return type is
  the same or a superclass of the child method's return type.
* Only replace a child method with an equivalent parent method if
  the parent method's class is public.

This fixes #6197