Restructure multi-arg system with chdir to use exec script
This change helps make our sh-based chdir work properly without
us mashing the command + arguments together into a string that may
parse badly when passed to sh -c. The primary trick here is to use
sh to do the chdir followed by an exec of the explicit argument
list passed in. The chdir will run in the intermediate shell, as
with a fork/exec, and then the exec will receive the actual target
program and its arguments unpolluted.

With this change, the example from #6153 works properly, and the
Ruby script is passed along without sh seeing it as part of a
shell script.