Avoid binding Kotlin Companion classes
Kotlin's companions are implemented similarly to Scala singletons,
except that both a Companion inner class and static final field
are generated. As a result, we warn when trying to redefine that
Companion constant. This is the cause of the warning in #6196.

The fix here avoids binding the inner class, since it's likely not
intended to be visible to non-Kotlin consumers, and because the
actual instance of it seems like the more appropriate value to
bind.

Fixes #6196