Try to sync Java and Ruby exception causes
The two were completely separate before this, but now we attempt
to keep them in sync:

* If the Java side calls initCause, we set that cause into the
  Ruby side.
* If the cause is set in the Ruby side, we will attempt to call
  initCause on the Java side iff it has not already been
  initialized

The lifecycle of cause in Ruby is a bit different from Java, with
a few different ways it can be set or "forced", so we can't make
it a single-shot initialization like with Throwable.initCause.

This relates to #6203.