Try to make the package module more "real"
Our package modules are weirdly structured to try to have as blank
a method table as possible, but this ends up producing objects
that look like classes or modules but do not behave as such. This
commit removes some of the sneakier bits of this structure and
makes Java package modules simply extend Module, so they do not
trip up reflective tools like Sorbet as reported in #6242.

This change may need further refinement to ensure that valid
package names shadowed by Module methods can still be accessed.