Try to estimate size of interpolated string
The heuristic here adds up the actual length of string elements
plus the estimated length of some other elements (defaulting to a
minimum of 4) and uses that amount * 1.5 to give buffer room when
interpolating. The largest cost of interpolating a string seems to
be when the initial size is too low and it must be repeatedly
resized. This is a "dumb" static estimate that could be improved
in the JIT by using indy and tracking string sizes that result
from a given dstr.

With this change, a benchmark of trivial string interpolation is
now as fast or faster than JRuby 1.7.