Eagerly compile methods which contain break/next.
Ruby expest invalid break and next keywords to raise syntax
errors when not in the right place but we lazily build methods.
Solution is to mark methods which contain next/break and eagerly
build those even though we may not compile them.  This will have
some memory hit but most methods to not contain them so it should
be very small.