Treat 8bit strings as raw bytes in expand_path
In order to be able to expand paths that are marked as binary, we
need to be able to treat those paths as raw bytes. This change
attempts to decode those paths as ISO-8859-1 bytes, allowing the
path expansion to ignore any multibyte characters rather than
improperly decoding them. The final encoding is used to put the
ISO-8859-1 bytes back into their 8-bit form and re-mark them as
the eventually negotiated encoding.

This fixes part of the WEBrick "cjk" test failure by allowing the
binary-encoded multibyte path to be expanded without mangling the
MBC character.

This change is limited to when the incoming strings are explicitly
marked as "binary" using the ASCII-8BIT encoding, since that case
clearly has no encoding hint for us to use to get characters.
There may be other cases, like CR_UNKNOWN or CR_BROKEN, that also
deserve this treatment, but for now the change is limited to
explicitly binary strings.

See #6171 for the general WEBRick bug and #6246 for the PR that
attempts to fix failures.