Treat binary strings as system encoding
Our current logic for dealing with file existence is intertwined
with a lot of JDK file APIs, so we don't have the option here
of just using a raw-encoded string. This change attempts to use
the system default encoding when the incoming path is specified as
"binary" so that hopefully its characters will be decoded
properly.

This fixes cases where a properly-encoded multibyte path is
marked as binary, as is the case in the WEBrick file-serving
pipeline described in #6246.