Raise a NotFound error when opening a non-existent URL
This isn't great since we're allowing a null field and checking
for null, but the logic surrounding this has been designed to
lazily create the channel or stream in question. This is likely
due to the use of URLResource and its siblings as generic path/URI
mungers that can clean up and canonicalize/abosolutize even paths
that do not exist. Rather than attempt to isolate that behavior
(which might be a lot more work), I've opted to mimic FileResource
in raising an error along the lazy path when it's clear we don't
have a resource in hand to open.

This fixes #6219.