Elininate MainTestSuite and run individual cases
MainTestSuite includes and somewhat hides a large number of test
suites, resulting a large percentage of our JUnit test run time
being spent silent. This change expands the test target to run
each of the MainTestSuite test cases individually, so they show
up in logs.

This also includes a rename of TestRubyBase, a base class for
several test cases, to Base. The original name was being picked
up as a test class, and then failing because it had no test
methods.