Fixes much more than a single tag but that tag was several tests.
Also refactored all error strings to be sourced out of the parser.
The new format add {{whatever}} as the tokens yyName in our generated
Java code.  This has been a long standing issue which is not super
important but does allow us to get complete parity with MRI
warning and error messages coming out of the parser.