Expose Java's built-in monitor synchronization as a JRuby API
I'm not sure why we've never done this before, but since there's
really no way to synchronize against an object's Java monitor from
Ruby this API seems like a gap in JRuby. Here I've added the
following JRuby utility functions:

* JRuby.synchronized(arg) {} which works like Java's synchronized
  statement
* JRuby.wait(arg [, millis [, nanos]]) like Object.wait
* JRuby.notify(arg) and notifyAll(arg) like Object.notify[All]

This question came up in the JRuby chat, and it seems like a
useful and appropriate addition to the JRuby module.