Always encode gethostname result as UTF-8
Our gethostname backends on the JDK's socket API, which always
returns a decoded character string. Rather than try to pretend
that this decoding has not happened and stuff the string back into
a "binary" ASCII-8BIT string, we instead always normalize the
encoding to UTF-8, the default encoding for Ruby strings now.

Fixes #6156