Don't initialize cause if null/nil
The improvements in #6234 exposed a flaw in how we construct the
Throwable associated with a ruby exception. Specifically, by
requesting the Throwable instance, we pass through
RaiseException's constructor, which triggers a call to `preRais`
that populates the backtrace at that point. When this happens
during a call to `Thread#raise` it ends up propagating an
exception object with the backtrace of the caller rather than of
the thread. This led to the following failures, which depend on
the backtrace being the one from the thread:

```
TestIO#test_race_closed_stream:
IOError: stream closed in another thread
    org/jruby/RubyThread.java:1374:in `raise'
    org/jruby/RubyIO.java:1990:in `close'
    /home/travis/build/jruby/jruby/test/mri/ruby/test_io.rb:3622:in `block in <main>'
    org/jruby/RubyIO.java:4197:in `pipe'
    /home/travis/build/jruby/jruby/test/mri/ruby/test_io.rb:3608:in `<main>'
    /home/travis/build/jruby/jruby/test/mri/ruby/test_io.rb:3603:in `test_race_closed_stream'

Thread#raise on a sleeping thread raises the given exception and the backtrace is the one of the interrupted thread FAILED
Expected "org/jruby/RubyThread.java:1392:in `raise'" to include "sleep"
/home/travis/build/jruby/jruby/spec/ruby/core/thread/raise_spec.rb:49:in `block in <main>'
org/jruby/RubyBasicObject.java:2695:in `instance_exec'
org/jruby/RubyArray.java:4555:in `all?'
org/jruby/RubyArray.java:1815:in `each'
/home/travis/build/jruby/jruby/spec/ruby/core/thread/raise_spec.rb:14:in `<main>'
org/jruby/RubyKernel.java:1078:in `load'
org/jruby/RubyBasicObject.java:2695:in `instance_exec'
org/jruby/RubyArray.java:1815:in `each'
```

This commit dodges the issue by not attempting to set the cause
if it is null or nil, since this would be either a no-op or an
error on the Throwable (Java cause can only be initialized once).
The root cause is filed as #6277.