Always use 9.2.9.0, the last to support --no-rdoc
This is a partial fix for #6103. The issue there is that in 9.2.10
we updated to a version of RubyGems that no longer supports the
--no-rdoc and --no-ri flags. Normally the build here would run
with 9.2.7.0, which does support those flags, but when running a
SNAPSHOT build we use the version of JRuby that we're building.
If that version is >= 9.2.10.0 it will fail because the --no-rdoc
flag is not recognized.

We could fix it by enabling rdoc and bypassing the relevant logic
in the InstallMojo from gem-maven-plugin, but we don't really want
to install rdoc and ri for the bootstrap gems.

Instead, I am forcing this to always use 9.2.7.0 until we can get
the plugin updated to a newer JRuby and fix the flags
appropriately.

This allows a non-SNAPSHOT bootstrap to almost work; it then fails
with the issue in #6174. I will commit a separate workaround for
that.

Works around #6103.