Don't install the jruby-launcher during bootstrap
This gem was only being installed for non-snapshot (release)
builds, even though we don't ever ship the results of that
install. Because of #6174, that install fails because no
jruby.home is passed in the extension-building subprocess.

This works around the issue by not installing jruby-launcher for
release builds, but the problem remains: gem-maven-plugin cannot
currently install gems with native builds due to the build command
line not providing a jruby.home.

This relates to #6103, which prevented us from getting to this
point before if building a non-snapshot build. With this change,
both snapshots and releases will bootstrap correctly, and neither
will install jruby-launcher.