Skip invasive test and add one that works on Java 9+
The skipped test will fail on Java 9+ because we are calling from
a named module into a java.base package that's not open. In order
to keep the basic spirit of the test running, I added a similar
test against the JI fixtures we load via classpath.

These APIs (JavaClass etc) are deprecaed, but we do not yet have
a clean path to replacing and removing them. Therefore we should
at least keep basic tests working.