Rework exception translation for more types
This expands our exception translation to work for more than just
the IOException family of types, first including the
NotYetConnectedException that does not live within that hierarchy.

In addition, it makes this translation method explicitly from
Throwable to Throwable with a user-provided fallback, and adds a
method for doing this translation or throwing the original error
in one shot.

This change was motivated by issues in BasicSocket where the
NotYetConnectedException was bubbling out as-is, rather than being
translated into an ENOTCONN. A subsequent change will modify
RubyBasicSocket to use this new logic.