Replace nearly all SizeFn impls and instances with method refs
This eliminates the class for all remaining SizeFn inner class
implementations and replaces all stateful inner class and lambda
implementations of SizeFn with method references. This completes
all requirements for #4688.

Method references are zero-alloc after the initial access, and
are initialized lazily as opposed to anonymous inner classes
(allocated every time) or static impl references (initialized at
class init time). In addition, through the magic of lambdas this
eliminates any remaining inner class instances, shrinking our
class load slightly.

The only controversial change here is the generification of SizeFn
to allow the passed-in self to be of the same type as the object
that serves as the enunmeration base. This might introduce some
binary incompatibility with external libraries, but I do not know
of any such libraries that implement their own SizeFn.