FIC rules the pass universe now.
Compiler passes would access all data via IRScope.  fullinterpretercontext stores
info needed by passes and it can potentially race in cases where the same scope
is compiled at the same time.  Most of this change worked out well because much
of the data needed by passes is stored in fic already.  The remaining missing
pieces is:

 1. flags
 2. temp variabels
 3. instrs storing scope vs ic

In truth, #3 may not be a problem but it is added for consideration.

The second thing this commit does is remove buildComplete().  Now when you
prepareFullBuild or prepareForCompilation is fully build it and then replace
IRScopes fullIntperpreterContext field at that point.  No partial full build
checks.  This works because the passes now pass around the in-flight fic vs
constantly asking IRScope for it.