Propagate code range into Symbol and calculate hash like String
This should finally fix all remaining issues with Symbol encodings
not reflecting the original encoding they were interned with.

In #1348 a case was reported where 7-bit ASCII bytes forced to
UTF-16 did not intern separately from the same bytes as US-ASCII.
This was mostly due to the hash calculation for the Symbol table
not considering the String's encoding, in the way that String#hash
does.

This patch makes the following changes:

* Propagate code range into RubySymbol and make it a CodeRangeable
* Use the same hash logic for Symbol's ByteList + CR that String
  uses.
* Fix lookup and construction paths that did not provide CR to
  scan for it.
* Modify paths that have CR available to pass it in, avoiding the
  scan.
* Move generic CodeRangeable logic to the interface as default
  methods.