Flagocaplypse.
Most flags are toggle on and booleans on IRScope.  None of these could possibly be changed
back to false with how we run JRuby today.  The inliner and presence of speculative inlining
could change some, but when that point comes we will have redesigned how we cache state about
what we are compiling.

Partially, this patch does isolate the things which are still in flags between IRBuilder and
FullInterpreterContext.  Because IRScope booleans are set by startup time and the rest are
contained within FIC there should be no ability for two in-flight FICs to stomp on each others
state.