Eliminate "fast" but wrong j.l.String to Symbol logic
With the work in #6341 to properly hash symbols using bytes and
encoding, the previous logic that could do a "fast" retrieval of
a Symbol for a given Java String is now incorrect. The logic to
retrieve a Symbol must always match the String hash logic, which
requires the encoded bytes plus the encoding, with an optional
code range to avoid doing a scan.

The "fast" path to get a Symbol is now via passing a ByteList and
code range int, or optionally omitting the code range and
accepting that it will be scanned for.