Clean up scope-searching logic
Previous changes marked lambdas as return target when looking for
a target scope, but did not consider it a target for searching the
stack of active scopes. This should fix the resulting regression.

Note that the logic for determining if a DynamicScope is a return
target has now been moved into DynamicScope, and the double-check
in checkForLJE which caused the regression has been removed. That
method calls getContainingReturnToScope which already guarantees
the resulting scope is a valid return target.