Use sun.stdout.encoding property for console
The System.console + reflection method of getting the console
encoding is unreliable, since if stdio has been connected to a
pipe the System.console will be null, causing us to fall back on
an incorrect console encoding. This adds logic to first try the
sun.stdout.encoding property, which is populated using the same
Win32 GetConsoleCP function as the Console encoding. This property
if present should allow us to always get the proper encoding for
the console, regardless of whether stdio is a pipe or the actual
console.