Add support for TCP_CORK (experimental)
On Linux, there are issues when benchmarking network requests with
very small packet sizes due to the delayed ACK problem. In this
mode, the extra ACK sent to acknowledge packets are frequently
delayed until there's real data to be sent by the server, and then
piggy-back on those data packets. This avoids sending a unique ACK
for every packet and improves performance on a chatty connection.
However in some cases, this causes some packets to wait up to 40ms
for an ACK, which causes small benchmarks of servers like Puma to
run very slowly with very little CPU usage.

It is possible to specify TCP_NODELAY to disable this behavior,
but this flag will be removed from the socket for many nnormal
operations and must be set periodically on both the client side
and the server side, which makes it a nonstarter when the client
is out of our control.

Instead, servers like Puma may use the Linux-specific TCP_CORK
mode, which "bottles" data packets until they are too large to
buffer, then sending them together. Puma implements this flag in
the following code:

https://github.com/puma/puma/blob/b981e97b81c6dc288081cee5ca6a2a603c047e78/lib/puma/server.rb#L101-L124

Puma calls these methods before and after handling a request, to
"cork" the data and then release it, which improves latency of
small request/response cycles.

https://github.com/puma/puma/blob/b981e97b81c6dc288081cee5ca6a2a603c047e78/lib/puma/server.rb#L642-L772

A longer explanation of TCP_CORK and the delayed ACK problem can
be read here: https://baus.net/on-tcp_cork/

This patch adds experimental support for using the native
setsockopt against a Java socket, so that we can pass through the
TCP_CORK option not supported by the JDK. Puma will need an
additional patch to not use RUBY_PLATFORM to detect Linux, since
that constant always points to "java" for us.